fix: sanitize quote response amount
I realized that we build quote responses with the _specified_ swap
amount returned by the quoter as well as the unspecified one. This could
cause downstream bugs if we assume the values returned by quoter are the
ones expected by swapper. this commit just munches the specified amount
to be the requested one from the swapper